Remove ENABLE_CAS3V2_API flag
This removes all branches around the feature flag
