+++ helm_deploy/values-test.yaml
@@ -24,7 +24,6 @@ generic-service:
     DOMAIN_EVENTS_EMIT_ENABLED: 'personArrived,personDeparted'
     CANCEL_SCHEDULED_ARCHIVE_ENABLED: true
     SHOW_GAP_REPORT_BUTTON: true
-    ENABLE_CAS3V2_API: true
     IN_MAINTENANCE_MODE: false
 
   allowlist: null
