+++ integration_tests/mockApis/confirmation.ts
@@ -1,11 +1,10 @@
 import type { SuperAgentRequest } from 'superagent'
-import type { Cas3Confirmation, Confirmation } from '@approved-premises/api'
+import type { Cas3Confirmation } from '@approved-premises/api'
 
 import { getMatchingRequests, stubFor } from '.'
 import paths from '../../server/paths/api'
-import config from '../../server/config'
 
-const cas3v2ApiEnabledStubs = {
+export default {
   stubConfirmationCreate: (args: {
     premisesId: string
     bookingId: string
@@ -30,31 +29,3 @@ const cas3v2ApiEnabledStubs = {
       })
     ).body.requests,
 }
-
-const cas3v2ApiDisabledStubs = {
-  stubConfirmationCreate: (args: {
-    premisesId: string
-    bookingId: string
-    confirmation: Confirmation
-  }): SuperAgentRequest =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/confirmations`,
-      },
-      response: {
-        status: 201,
-        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
-        jsonBody: args.confirmation,
-      },
-    }),
-  verifyConfirmationCreate: async (args: { premisesId: string; bookingId: string }) =>
-    (
-      await getMatchingRequests({
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/confirmations`,
-      })
-    ).body.requests,
-}
-
-export default config.flags.enableCas3v2Api ? cas3v2ApiEnabledStubs : cas3v2ApiDisabledStubs
