+++ integration_tests/mockApis/departure.ts
@@ -1,14 +1,13 @@
 import type { Response, SuperAgentRequest } from 'superagent'
 
-import type { Cas3Departure, Departure } from '@approved-premises/api'
+import type { Cas3Departure } from '@approved-premises/api'
 
 import { getMatchingRequests, stubFor } from '.'
 import { errorStub } from './utils'
 import { departureReasons, moveOnCategories } from '../../server/testutils/stubs/referenceDataStubs'
 import paths from '../../server/paths/api'
-import config from '../../server/config'
 
-const cas3v2ApiEnabledStubs = {
+export default {
   stubDepartureCreate: (args: { premisesId: string; bookingId: string; departure: Cas3Departure }): SuperAgentRequest =>
     stubFor({
       request: {
@@ -39,31 +38,3 @@ const cas3v2ApiEnabledStubs = {
       })
     ).body.requests,
 }
-
-const cas3v2ApiDisabledStubs = {
-  stubDepartureCreate: (args: { premisesId: string; bookingId: string; departure: Departure }): SuperAgentRequest =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/cas3/premises/${args.premisesId}/bookings/${args.bookingId}/departures`,
-      },
-      response: {
-        status: 201,
-        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
-        jsonBody: args.departure,
-      },
-    }),
-  stubDepartureCreateErrors: (args: { premisesId: string; bookingId: string; params: Array<string> }) =>
-    stubFor(errorStub(args.params, `/cas3/premises/${args.premisesId}/bookings/${args.bookingId}/departures`, 'POST')),
-  stubDepartureReferenceData: (): Promise<[Response, Response]> =>
-    Promise.all([stubFor(departureReasons), stubFor(moveOnCategories)]),
-  verifyDepartureCreate: async (args: { premisesId: string; bookingId: string }) =>
-    (
-      await getMatchingRequests({
-        method: 'POST',
-        url: `/cas3/premises/${args.premisesId}/bookings/${args.bookingId}/departures`,
-      })
-    ).body.requests,
-}
-
-export default config.flags.enableCas3v2Api ? cas3v2ApiEnabledStubs : cas3v2ApiDisabledStubs
