+++ integration_tests/mockApis/extension.ts
@@ -1,13 +1,12 @@
 import { SuperAgentRequest } from 'superagent'
 
-import type { Cas3Extension, Extension } from '@approved-premises/api'
+import type { Cas3Extension } from '@approved-premises/api'
 
 import { getMatchingRequests, stubFor } from '.'
 import { bedspaceConflictResponseBody, errorStub } from './utils'
 import paths from '../../server/paths/api'
-import config from '../../server/config'
 
-const cas3v2ApiEnabledStubs = {
+export default {
   stubExtensionCreate: (args: { premisesId: string; bookingId: string; extension: Cas3Extension }): SuperAgentRequest =>
     stubFor({
       request: {
@@ -59,52 +58,3 @@ const cas3v2ApiEnabledStubs = {
       })
     ).body.requests,
 }
-
-const cas3v2ApiDisabledStubs = {
-  stubExtensionCreate: (args: { premisesId: string; bookingId: string; extension: Extension }): SuperAgentRequest =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/extensions`,
-      },
-      response: {
-        status: 201,
-        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
-        jsonBody: args.extension,
-      },
-    }),
-  stubExtensionCreateErrors: (args: {
-    premisesId: string
-    bookingId: string
-    params: Array<string>
-  }): SuperAgentRequest =>
-    stubFor(errorStub(args.params, `/premises/${args.premisesId}/bookings/${args.bookingId}/extensions`, 'POST')),
-  stubExtensionCreateConflictError: (args: {
-    premisesId: string
-    bookingId: string
-    conflictingEntityId: string
-    conflictingEntityType: 'booking' | 'lost-bed'
-  }) =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/extensions`,
-      },
-      response: {
-        status: 409,
-        headers: {
-          'Content-Type': 'application/problem+json;charset=UTF-8',
-        },
-        jsonBody: bedspaceConflictResponseBody(args.conflictingEntityId, args.conflictingEntityType),
-      },
-    }),
-  verifyExtensionCreate: async (args: { premisesId: string; bookingId: string }) =>
-    (
-      await getMatchingRequests({
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/extensions`,
-      })
-    ).body.requests,
-}
-
-export default config.flags.enableCas3v2Api ? cas3v2ApiEnabledStubs : cas3v2ApiDisabledStubs
