+++ integration_tests/mockApis/turnaround.ts
@@ -1,13 +1,12 @@
 import { SuperAgentRequest } from 'superagent'
 
-import type { Cas3Turnaround, Turnaround } from '@approved-premises/api'
+import type { Cas3Turnaround } from '@approved-premises/api'
 
 import { getMatchingRequests, stubFor } from '.'
 import { bedspaceConflictResponseBody, errorStub } from './utils'
 import paths from '../../server/paths/api'
-import config from '../../server/config'
 
-const cas3v2ApiEnabledStubs = {
+export default {
   stubTurnaroundCreate: (args: {
     premisesId: string
     bookingId: string
@@ -63,52 +62,3 @@ const cas3v2ApiEnabledStubs = {
       })
     ).body.requests,
 }
-
-const cas3v2ApiDisabledStubs = {
-  stubTurnaroundCreate: (args: { premisesId: string; bookingId: string; turnaround: Turnaround }): SuperAgentRequest =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/turnarounds`,
-      },
-      response: {
-        status: 201,
-        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
-        jsonBody: args.turnaround,
-      },
-    }),
-  stubTurnaroundCreateErrors: (args: {
-    premisesId: string
-    bookingId: string
-    params: Array<string>
-  }): SuperAgentRequest =>
-    stubFor(errorStub(args.params, `/premises/${args.premisesId}/bookings/${args.bookingId}/turnarounds`, 'POST')),
-  stubTurnaroundCreateConflictError: (args: {
-    premisesId: string
-    bookingId: string
-    conflictingEntityId: string
-    conflictingEntityType: 'booking' | 'lost-bed'
-  }) =>
-    stubFor({
-      request: {
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/turnarounds`,
-      },
-      response: {
-        status: 409,
-        headers: {
-          'Content-Type': 'application/problem+json;charset=UTF-8',
-        },
-        jsonBody: bedspaceConflictResponseBody(args.conflictingEntityId, args.conflictingEntityType),
-      },
-    }),
-  verifyTurnaroundCreate: async (args: { premisesId: string; bookingId: string }) =>
-    (
-      await getMatchingRequests({
-        method: 'POST',
-        url: `/premises/${args.premisesId}/bookings/${args.bookingId}/turnarounds`,
-      })
-    ).body.requests,
-}
-
-export default config.flags.enableCas3v2Api ? cas3v2ApiEnabledStubs : cas3v2ApiDisabledStubs
