+++ server/services/bookingSearchService.ts
@@ -1,12 +1,15 @@
-import type { BookingSearchApiStatus, BookingSearchParameters, TableRow } from '@approved-premises/ui'
+import type {
+  BookingSearchApiStatus,
+  BookingSearchParameters,
+  PaginatedResponse,
+  TableRow,
+} from '@approved-premises/ui'
 import { BookingSearchResult, Cas3BookingSearchResult } from '@approved-premises/api'
 import type { RestClientBuilder } from '../data'
 import BookingClient from '../data/bookingClient'
 import { CallConfig } from '../data/restClient'
 import paths from '../paths/temporary-accommodation/manage'
 import { DateFormats } from '../utils/dateUtils'
-import { PaginatedResponse } from '../@types/ui'
-import { bookingSearchResultsToCas3BookingSearchResults } from '../utils/bookingSearchUtils'
 
 export default class BookingSearchService {
   constructor(private readonly bookingClientFactory: RestClientBuilder<BookingClient>) {}
@@ -27,9 +30,7 @@ export default class BookingSearchService {
       totalPages,
       totalResults,
       url,
-      data: bookingSearchResultsToCas3BookingSearchResults(
-        data as Array<BookingSearchResult> | Array<Cas3BookingSearchResult>,
-      ).map(summary => {
+      data: data.map(summary => {
         return [
           this.textValue(summary.person.name || 'Limited access offender'),
           this.textValue(summary.person.crn),
