+++ server/services/premisesService.ts
@@ -8,7 +8,6 @@ import {
   Cas3ReferenceData,
   Cas3UnarchivePremises,
   Cas3UpdatePremises,
-  Characteristic,
   LocalAuthorityArea,
   ProbationDeliveryUnit,
   ProbationRegion,
@@ -16,9 +15,6 @@ import {
 import { PremisesClient, ReferenceDataClient, RestClientBuilder } from '../data'
 
 import { CallConfig } from '../data/restClient'
-import { characteristicToCas3ReferenceData, filterCharacteristics } from '../utils/characteristicUtils'
-import { populatePremisesCharacteristics } from '../utils/premisesUtils'
-import config from '../config'
 
 export type Cas3PremisesReferenceData = {
   localAuthorities: Array<LocalAuthorityArea>
@@ -45,7 +41,7 @@ export default class PremisesService {
 
   async getSinglePremises(callConfig: CallConfig, premisesId: string): Promise<Cas3Premises> {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.find(premisesId))
+    return premisesClient.find(premisesId)
   }
 
   async getSinglePremisesBedspaceTotals(
@@ -58,13 +54,6 @@ export default class PremisesService {
 
   async getCharacteristics(callConfig: CallConfig): Promise<Array<Cas3ReferenceData>> {
     const referenceDataClient = this.referenceDataClientFactory(callConfig)
-
-    if (!config.flags.enableCas3v2Api) {
-      return filterCharacteristics(
-        await referenceDataClient.getReferenceData<Characteristic>('characteristics'),
-        'premises',
-      ).map(characteristicToCas3ReferenceData)
-    }
 
     return referenceDataClient.getCas3ReferenceData('PREMISES_CHARACTERISTICS')
   }
@@ -95,12 +84,12 @@ export default class PremisesService {
 
   async createPremises(callConfig: CallConfig, newPremises: Cas3NewPremises): Promise<Cas3Premises> {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.create(newPremises))
+    return premisesClient.create(newPremises)
   }
 
   async updatePremises(callConfig: CallConfig, premisesId: string, updatedPremises: Cas3UpdatePremises) {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.update(premisesId, updatedPremises))
+    return premisesClient.update(premisesId, updatedPremises)
   }
 
   async canArchivePremises(callConfig: CallConfig, premisesId: string) {
@@ -110,21 +99,21 @@ export default class PremisesService {
 
   async archivePremises(callConfig: CallConfig, premisesId: string, archivePayload: Cas3ArchivePremises) {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.archive(premisesId, archivePayload))
+    return premisesClient.archive(premisesId, archivePayload)
   }
 
   async unarchivePremises(callConfig: CallConfig, premisesId: string, unarchivePayload: Cas3UnarchivePremises) {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.unarchive(premisesId, unarchivePayload))
+    return premisesClient.unarchive(premisesId, unarchivePayload)
   }
 
   async cancelArchivePremises(callConfig: CallConfig, premisesId: string) {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.cancelArchive(premisesId))
+    return premisesClient.cancelArchive(premisesId)
   }
 
   async cancelUnarchivePremises(callConfig: CallConfig, premisesId: string) {
     const premisesClient = this.premisesClientFactory(callConfig)
-    return populatePremisesCharacteristics(await premisesClient.cancelUnarchive(premisesId))
+    return premisesClient.cancelUnarchive(premisesId)
   }
 }
