+++ server/utils/bedspaceSearchUtils.test.ts
@@ -1,17 +1,10 @@
 import { ObjectWithDateParts } from '@approved-premises/ui'
-import {
-  cas3BedspaceSearchResultsToCas3v2BedspaceSearchResults,
-  characteristicPairsToCas3CharacteristicPairs,
-  isCas3v2BedspaceSearchResults,
-  overlapsToCas3v2BedspaceSearchResultOverlaps,
-  validateSearchQuery,
-} from './bedspaceSearchUtils'
+import { validateSearchQuery } from './bedspaceSearchUtils'
 import {
   bedspaceSearchResultsFactory,
   cas3v2BedspaceSearchResultsFactory,
   overlapFactory,
 } from '../testutils/factories'
-import cas3v2BedspaceSearchResults from '../testutils/factories/cas3v2BedspaceSearchResults'
 import referenceDataFactory from '../testutils/factories/referenceData'
 
 type BedspaceSearchQuery = ObjectWithDateParts<'startDate'> & {
@@ -110,88 +103,5 @@ describe('validateSearchQuery', () => {
   it('returns null when there are no validation errors', () => {
     const result = validateSearchQuery(query)
     expect(result).toBeNull()
-  })
-
-  // TODO -- ENABLE_CAS3V2_API cleanup: remove the following casting utilities tests
-  describe('Cas3v2 casting utilities', () => {
-    describe('isCas3v2BedspaceSearchResults', () => {
-      it('returns true for a Cas3v2BedspaceSearchResults', () => {
-        const results = cas3v2BedspaceSearchResultsFactory.build()
-
-        expect(isCas3v2BedspaceSearchResults(results)).toEqual(true)
-      })
-
-      it('returns false for a Cas3BedspaceSearchResults', () => {
-        const results = bedspaceSearchResultsFactory.build()
-
-        expect(isCas3v2BedspaceSearchResults(results)).toEqual(false)
-      })
-    })
-
-    describe('characteristicPairsToCas3CharacteristicPairs', () => {
-      it('transforms a list of CharacteristicPair to a list of Cas3CharacteristicPair', () => {
-        const characteristicPairs = referenceDataFactory.characteristic('room').buildList(1)
-
-        expect(characteristicPairsToCas3CharacteristicPairs(characteristicPairs)).toEqual([
-          { name: characteristicPairs[0].name, description: characteristicPairs[0].name },
-        ])
-      })
-    })
-
-    describe('overlapsToCas3v2BedspaceSearchResultOverlaps', () => {
-      it('transforms a list of Cas3BedspaceSearchResultOverlap to a list of Cas3v2BedspaceSearchResultOverlap', () => {
-        const overlap = overlapFactory.build()
-
-        expect(overlapsToCas3v2BedspaceSearchResultOverlaps([overlap])).toEqual([
-          {
-            assessmentId: overlap.assessmentId,
-            bedspaceId: overlap.roomId,
-            bookingId: overlap.bookingId,
-            crn: overlap.crn,
-            days: overlap.days,
-            isSexualRisk: overlap.isSexualRisk,
-            name: overlap.name,
-            sex: overlap.sex,
-            personType: overlap.personType,
-          },
-        ])
-      })
-    })
-
-    describe('cas3BedspaceSearchResultsToCas3v2BedspaceSearchResults', () => {
-      it('returns a Cas3v2BedspaceSearchResults directly', () => {
-        const searchResults = cas3v2BedspaceSearchResults.build()
-
-        const results = cas3BedspaceSearchResultsToCas3v2BedspaceSearchResults(searchResults)
-
-        expect(isCas3v2BedspaceSearchResults(results)).toEqual(true)
-        expect(results).toEqual(searchResults)
-      })
-
-      it('transforms a Cas3BedspaceSearchResults to a Cas3v2BedspaceSearchResults', () => {
-        const searchResults = bedspaceSearchResultsFactory.build()
-
-        const results = cas3BedspaceSearchResultsToCas3v2BedspaceSearchResults(searchResults)
-
-        expect(isCas3v2BedspaceSearchResults(results)).toEqual(true)
-        expect(results.results[0]).toEqual({
-          premises: expect.objectContaining({
-            bedspaceCount: searchResults.results[0].premises.bedCount,
-            bookedBedspaceCount: searchResults.results[0].premises.bookedBedCount,
-            characteristics: characteristicPairsToCas3CharacteristicPairs(
-              searchResults.results[0].premises.characteristics,
-            ),
-          }),
-          bedspace: expect.objectContaining({
-            id: searchResults.results[0].bed.id,
-            reference: searchResults.results[0].room.name,
-            characteristics: characteristicPairsToCas3CharacteristicPairs(
-              searchResults.results[0].room.characteristics,
-            ),
-          }),
-          overlaps: overlapsToCas3v2BedspaceSearchResultOverlaps(searchResults.results[0].overlaps),
-        })
-      })
-    })
   })
 })
