+++ server/utils/bedspaceUtils.ts
@@ -265,19 +265,3 @@ const htmlValue = (value: string) => {
 }
 
 const formatNotes = (notes: string): string => notes.replace(/\n/g, '<br />')
-
-// TODO -- ENABLE_CAS3V2_API cleanup: remove the following casting utilities and all usages
-export const characteristicToBedspaceCharacteristic = (characteristic: Characteristic): Cas3BedspaceCharacteristic => ({
-  id: characteristic.id,
-  description: characteristic.name,
-  name: characteristic.propertyName,
-})
-
-export const populateBedspaceCharacteristics = (bedspace: Cas3Bedspace): Cas3Bedspace => {
-  if (bedspace.bedspaceCharacteristics) return bedspace
-
-  return {
-    ...bedspace,
-    bedspaceCharacteristics: bedspace.characteristics.map(characteristicToBedspaceCharacteristic),
-  }
-}
