+++ server/utils/characteristicUtils.ts
@@ -1,4 +1,4 @@
-import { Cas3ReferenceData, Characteristic } from '../@types/shared'
+import { Characteristic } from '@approved-premises/api'
 
 export const filterCharacteristics = (
   characteristics: Array<Characteristic>,
@@ -8,10 +8,3 @@ export const filterCharacteristics = (
     characteristic => characteristic.modelScope === scope || characteristic.modelScope === '*',
   )
 }
-
-// TODO -- ENABLE_CAS3V2_API cleanup: remove the following casting utility and all usages
-export const characteristicToCas3ReferenceData = (characteristic: Characteristic): Cas3ReferenceData => ({
-  id: characteristic.id,
-  description: characteristic.name,
-  name: characteristic.propertyName,
-})
