+++ server/utils/lostBedUtils.ts
@@ -1,4 +1,4 @@
-import type { Cas3VoidBedspace, Cas3VoidBedspaceStatus, LostBed } from '@approved-premises/api'
+import type { Cas3VoidBedspace, Cas3VoidBedspaceStatus } from '@approved-premises/api'
 import type { PageHeadingBarItem } from '@approved-premises/ui'
 import paths from '../paths/temporary-accommodation/manage'
 
@@ -43,25 +43,4 @@ export function lostBedActions(
     ]
   }
   return null
-}
-
-// TODO -- ENABLE_CAS3V2_API cleanup: remove the following casting utilities and all usages
-export const isCas3VoidBedspace = (lostBed: LostBed | Cas3VoidBedspace): lostBed is Cas3VoidBedspace =>
-  Boolean((lostBed as Cas3VoidBedspace).bedspaceId)
-
-export const lostBedToCas3VoidBedspace = (lostBed: LostBed | Cas3VoidBedspace): Cas3VoidBedspace => {
-  if (isCas3VoidBedspace(lostBed)) return lostBed
-
-  const { bedId, bedName, cancellation, costCentre, reason, status, ...sharedProperties } = lostBed
-
-  return {
-    ...sharedProperties,
-    bedspaceId: bedId,
-    bedspaceName: bedName,
-    cancellationDate: cancellation?.createdAt,
-    cancellationNotes: cancellation?.notes,
-    costCentre,
-    reason,
-    status,
-  }
 }
